# support 1 shot allreduce  in 1-node and 2-node using mscclpp