@Alcanderian 2025-05-22
The benchmark result seems not matching the given graph in mscclpp ![img](https://github.com/microsoft/mscclpp/blob/main/docs/figs/mscclpp_vs_nccl_comparison_num_nodes_2.jpeg?raw=true)

And I found that we may have to support two algo to cover large batch size decode.
https://github.com/microsoft/mscclpp/blob/main/python/mscclpp_benchmark/allreduce_bench.py#L182
```
        if memory.nbytes < 2**22:
            mscclpp_algos = [MscclppAllReduce5(mscclpp_group, memory, memory_out, N_GPUS_PER_NODE, proxy_service)]
        else:
            mscclpp_algos = [MscclppAllReduce4(mscclpp_group, memory, N_GPUS_PER_NODE, proxy_service)]
```