@zyksir 2025-05-23
> I found that we may have to support two algo to cover large batch size decode. https://github.com/microsoft/mscclpp/blob/main/python/mscclpp_benchmark/allreduce_bench.py#L182
> 
> ```
>         if memory.nbytes < 2**22:
>             mscclpp_algos = [MscclppAllReduce5(mscclpp_group, memory, memory_out, N_GPUS_PER_NODE, proxy_service)]
>         else:
>             mscclpp_algos = [MscclppAllReduce4(mscclpp_group, memory, N_GPUS_PER_NODE, proxy_service)]
> ```

![image](https://github.com/user-attachments/assets/4e35c9d4-e1f3-4641-a0bb-c63c26753fa7)
I found that even in the benchmark script provided by mscclpp, the performance on large message is slower than NCCL. So for now I am working on improvement on small message sizes.