@zyksir 2025-05-29
> Btw, by two-shot do you mean intra-node all-reduce + inter-node all-reduce? I think NCCL doesn't do hierarchical all-reduce yet?

The 2-node allreduce is actually intra-node reduce-scatter + inter-node send-recv + intra-node broadcast. I am not sure how NCCL does allreduce. but I think this is a different way.